J4mFx_app/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-channels==0.7.0
django-crispy-forms==1.11.2
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.8.6
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
